Restructured directories and namespaces
