+++ composer.json
@@ -34,7 +34,7 @@
     },
     "require": {
         "php": "^5.6 || ^7.0",
-        "doctrine/doctrine-module": " ^2.1.8",
+        "doctrine/doctrine-module": " ^3.0 || ^4.0",
         "laminas/laminas-hydrator": "^3.0",
         "laminas/laminas-modulemanager": "^2.7.2",
         "laminas/laminas-mvc": "^3.0.2",
@@ -42,7 +42,6 @@
         "laminas/laminas-zendframework-bridge": "^1.0"
     },
     "require-dev": {
-        "doctrine/doctrine-orm-module": "^1.1 || ^2.1.3",
         "laminas-api-tools/api-tools-provider": "^1.2",
         "laminas/laminas-coding-standard": "~1.0.0",
         "laminas/laminas-i18n": "^2.7.3",
