+++ src/Filter/GreaterThan.php
@@ -6,15 +6,10 @@
  * @license   https://github.com/laminas-api-tools/api-tools-doctrine-querybuilder/blob/master/LICENSE.md New BSD License
  */
 
-namespace Laminas\ApiTools\Doctrine\QueryBuilder\Filter\ORM;
+namespace Laminas\ApiTools\Doctrine\QueryBuilder\Filter;
 
-class IsMemberOf extends AbstractFilter
+class GreaterThan extends AbstractFilter
 {
-    /**
-     * @param \Doctrine\ORM\QueryBuilder $queryBuilder
-     * @param $metadata
-     * @param $option
-     */
     public function filter($queryBuilder, $metadata, $option)
     {
         if (isset($option['where'])) {
@@ -33,10 +28,7 @@ class IsMemberOf extends AbstractFilter
             $option['alias'] = 'row';
         }
 
-        $format = null;
-        if (isset($option['format'])) {
-            $format = $option['format'];
-        }
+        $format = isset($option['format']) ? $option['format'] : null;
 
         $value = $this->typeCastField($metadata, $option['field'], $option['value'], $format);
 
@@ -44,7 +36,7 @@ class IsMemberOf extends AbstractFilter
         $queryBuilder->$queryType(
             $queryBuilder
                 ->expr()
-                ->isMemberOf(':' . $parameter, $option['alias'] . '.' . $option['field'])
+                ->gt($option['alias'] . '.' . $option['field'], ':' . $parameter)
         );
         $queryBuilder->setParameter($parameter, $value);
     }
