+++ src/Filter/IsNull.php
@@ -6,9 +6,9 @@
  * @license   https://github.com/laminas-api-tools/api-tools-doctrine-querybuilder/blob/master/LICENSE.md New BSD License
  */
 
-namespace Laminas\ApiTools\Doctrine\QueryBuilder\Filter\ORM;
+namespace Laminas\ApiTools\Doctrine\QueryBuilder\Filter;
 
-class LessThanOrEquals extends AbstractFilter
+class IsNull extends AbstractFilter
 {
     public function filter($queryBuilder, $metadata, $option)
     {
@@ -28,16 +28,10 @@ class LessThanOrEquals extends AbstractFilter
             $option['alias'] = 'row';
         }
 
-        $format = isset($option['format']) ? $option['format'] : null;
-
-        $value = $this->typeCastField($metadata, $option['field'], $option['value'], $format);
-
-        $parameter = uniqid('a');
         $queryBuilder->$queryType(
             $queryBuilder
                 ->expr()
-                ->lte($option['alias'] . '.' . $option['field'], ':' . $parameter)
+                ->isNull($option['alias'] . '.' . $option['field'])
         );
-        $queryBuilder->setParameter($parameter, $value);
     }
 }
