+++ src/Filter/OrX.php
@@ -6,11 +6,11 @@
  * @license   https://github.com/laminas-api-tools/api-tools-doctrine-querybuilder/blob/master/LICENSE.md New BSD License
  */
 
-namespace Laminas\ApiTools\Doctrine\QueryBuilder\Filter\ORM;
+namespace Laminas\ApiTools\Doctrine\QueryBuilder\Filter;
 
 use Doctrine\Common\Collections\ArrayCollection;
 
-class AndX extends AbstractFilter
+class OrX extends AbstractFilter
 {
     public function filter($queryBuilder, $metadata, $option)
     {
@@ -26,20 +26,21 @@ class AndX extends AbstractFilter
             $queryType = 'andWhere';
         }
 
-        $andX = $queryBuilder->expr()->andX();
-        $em   = $queryBuilder->getEntityManager();
-        $qb   = $em->createQueryBuilder();
+        $orX = $queryBuilder->expr()->orX();
+        $em  = $queryBuilder->getEntityManager();
+        $qb  = $em->createQueryBuilder();
 
         foreach ($option['conditions'] as $condition) {
-            $filter = $this->getFilterManager()->get(
-                strtolower($condition['type']),
-                [$this->getFilterManager()]
-            );
+            $filter = $this->getFilterManager()
+                ->get(
+                    strtolower($condition['type']),
+                    [$this->getFilterManager()]
+                );
             $filter->filter($qb, $metadata, $condition);
         }
 
         $dqlParts = $qb->getDqlParts();
-        $andX->addMultiple($dqlParts['where']->getParts());
+        $orX->addMultiple($dqlParts['where']->getParts());
         $queryBuilder->setParameters(
             new ArrayCollection(array_merge_recursive(
                 $queryBuilder->getParameters()->toArray(),
@@ -47,6 +48,6 @@ class AndX extends AbstractFilter
             ))
         );
 
-        $queryBuilder->$queryType($andX);
+        $queryBuilder->$queryType($orX);
     }
 }
