+++ src/OrderbyManagerFactory.php
@@ -6,11 +6,11 @@
  * @license   https://github.com/laminas-api-tools/api-tools-doctrine-querybuilder/blob/master/LICENSE.md New BSD License
  */
 
-namespace Laminas\ApiTools\Doctrine\QueryBuilder\Filter\Service;
+namespace Laminas\ApiTools\Doctrine\QueryBuilder;
 
 use Laminas\Mvc\Service\AbstractPluginManagerFactory;
 
-class ORMFilterManagerFactory extends AbstractPluginManagerFactory
+class OrderByManagerFactory extends AbstractPluginManagerFactory
 {
-    const PLUGIN_MANAGER_CLASS = ORMFilterManager::class;
+    const PLUGIN_MANAGER_CLASS = OrderByManager::class;
 }
