+++ test/assets/module/Db/config/module.config.php
@@ -2,45 +2,45 @@
 
 namespace Db;
 
-use Doctrine\ORM\Mapping\Driver\XmlDriver;
+use Doctrine\Mapping\Driver\XmlDriver;
 use Laminas\ApiTools\Doctrine\QueryBuilder\Filter;
 use Laminas\ServiceManager\Factory\InvokableFactory;
 
 return [
     'api-tools-doctrine-querybuilder-filter-orm' => [
         'aliases' => [
-            'eq'        => Filter\ORM\Equals::class,
-            'neq'       => Filter\ORM\NotEquals::class,
-            'lt'        => Filter\ORM\LessThan::class,
-            'lte'       => Filter\ORM\LessThanOrEquals::class,
-            'gt'        => Filter\ORM\GreaterThan::class,
-            'gte'       => Filter\ORM\GreaterThanOrEquals::class,
-            'isnull'    => Filter\ORM\IsNull::class,
-            'isnotnull' => Filter\ORM\IsNotNull::class,
-            'in'        => Filter\ORM\In::class,
-            'notin'     => Filter\ORM\NotIn::class,
-            'between'   => Filter\ORM\Between::class,
-            'like'      => Filter\ORM\Like::class,
-            'notlike'   => Filter\ORM\NotLike::class,
-            'orx'       => Filter\ORM\OrX::class,
-            'andx'      => Filter\ORM\AndX::class,
+            'eq'        => Filter\Equals::class,
+            'neq'       => Filter\NotEquals::class,
+            'lt'        => Filter\LessThan::class,
+            'lte'       => Filter\LessThanOrEquals::class,
+            'gt'        => Filter\GreaterThan::class,
+            'gte'       => Filter\GreaterThanOrEquals::class,
+            'isnull'    => Filter\IsNull::class,
+            'isnotnull' => Filter\IsNotNull::class,
+            'in'        => Filter\In::class,
+            'notin'     => Filter\NotIn::class,
+            'between'   => Filter\Between::class,
+            'like'      => Filter\Like::class,
+            'notlike'   => Filter\NotLike::class,
+            'orx'       => Filter\OrX::class,
+            'andx'      => Filter\AndX::class,
         ],
         'factories' => [
-            Filter\ORM\Equals::class              => InvokableFactory::class,
-            Filter\ORM\NotEquals::class           => InvokableFactory::class,
-            Filter\ORM\LessThan::class            => InvokableFactory::class,
-            Filter\ORM\LessThanOrEquals::class    => InvokableFactory::class,
-            Filter\ORM\GreaterThan::class         => InvokableFactory::class,
-            Filter\ORM\GreaterThanOrEquals::class => InvokableFactory::class,
-            Filter\ORM\IsNull::class              => InvokableFactory::class,
-            Filter\ORM\IsNotNull::class           => InvokableFactory::class,
-            Filter\ORM\In::class                  => InvokableFactory::class,
-            Filter\ORM\NotIn::class               => InvokableFactory::class,
-            Filter\ORM\Between::class             => InvokableFactory::class,
-            Filter\ORM\Like::class                => InvokableFactory::class,
-            Filter\ORM\NotLike::class             => InvokableFactory::class,
-            Filter\ORM\OrX::class                 => InvokableFactory::class,
-            Filter\ORM\AndX::class                => InvokableFactory::class,
+            Filter\Equals::class              => InvokableFactory::class,
+            Filter\NotEquals::class           => InvokableFactory::class,
+            Filter\LessThan::class            => InvokableFactory::class,
+            Filter\LessThanOrEquals::class    => InvokableFactory::class,
+            Filter\GreaterThan::class         => InvokableFactory::class,
+            Filter\GreaterThanOrEquals::class => InvokableFactory::class,
+            Filter\IsNull::class              => InvokableFactory::class,
+            Filter\IsNotNull::class           => InvokableFactory::class,
+            Filter\In::class                  => InvokableFactory::class,
+            Filter\NotIn::class               => InvokableFactory::class,
+            Filter\Between::class             => InvokableFactory::class,
+            Filter\Like::class                => InvokableFactory::class,
+            Filter\NotLike::class             => InvokableFactory::class,
+            Filter\OrX::class                 => InvokableFactory::class,
+            Filter\AndX::class                => InvokableFactory::class,
         ],
     ],
     'doctrine' => [
